chore(model): RPC conversion code isolation (O-R)
