+++ crates/api-model/src/rpc_conv/mod.rs
@@ -27,6 +27,17 @@ pub mod dhcp_record;
 pub mod dns;
 pub mod dpa_interface;
 pub mod dpu_remediation;
+pub mod operating_system_definition;
+pub mod os;
+pub mod power_manager;
+pub mod power_shelf;
+pub mod pxe;
+pub mod rack;
+pub mod rack_firmware;
+pub mod rack_type;
+pub mod redfish;
+pub mod resource_pool;
+pub mod route_server;
 pub mod site_explorer;
 pub mod sku;
 pub mod state_history;
